// src/Model/prodduct.ts
export interface Product {
    id:string
    productName: string; // Product name
    productDescription: string; // Description of the product
    productCode: string; // Product code generated by getRandomString
    productCategory: string; // Category of the product
    productPrice: number; // Product price parsed to integer
    quantity: number; // Quantity of the product parsed to integer
    productDiscount: number;// Indicates if the product has a discount
    productImage: string[]; // Product image, potentially an ID or index
    isActive: boolean; // Is the product active
    isOutOfStock: boolean; // Is the product out of stock
    createdAt: any // Timestamp of product creation
  }